outsource melscale to different file
